Move ComposedSection
